Rename store to localStore
