+++ packages/extension/src/app/cookieStore/index.tsx
@@ -27,7 +27,7 @@ import React, {
 /**
  * Internal dependencies.
  */
-import { type StorageValue, CookieStore, emptyTabData } from '../../store';
+import { type StorageValue, CookieStore, emptyTabData } from '../../localStore';
 import { getCurrentTabId } from '../../utils/getCurrentTabId';
 
 export interface ICookieStoreContext {
