+++ packages/extension/src/localStore/parseCookieHeader.ts
