+++ packages/extension/src/localStore/types.ts
