+++ packages/extension/src/localStore/updateStorage.ts
